srv/extraction-server/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
pytesseract
Pillow
python-magic
opencv-python
numpy
PyPDF2